+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Support `dropRole` command *[Feature]*
 * Fix concurrent upsert behavior, update the documents in case of conflicts during insert *[Bugfix]* (#295).
 * Support collation with `$sortArray` aggregation operator *[Feature]*
+* Add support for keyword `required` in `$jsonSchema`
 
 ### documentdb v0.107-0 (Unreleased) ###
 * Support sort by _id against the _id index using the enableIndexOrderbyPushdown flag *[Feature]*.